+++ libs/jacdac/jacdac.ts
@@ -388,7 +388,7 @@ namespace jacdac {
         StopStream
     }
 
-    function bufferEqual(l: Buffer, r: Buffer): boolean {
+    export function bufferEqual(l: Buffer, r: Buffer): boolean {
         if (!l || !r) return !!l == !!r;
         if (l.length != r.length) return false;
         for (let i = 0; i < l.length; ++i) {
@@ -477,7 +477,7 @@ namespace jacdac {
                         if (this.isConnected
                             && (!this._sendState
                                 || (control.millis() - this._sendTime > SensorHostDriver.MAX_SILENCE)
-                                || !bufferEqual(state, this._sendState))) {
+                                || !jacdac.bufferEqual(state, this._sendState))) {
 
                             // send state and record time
                             const pkt = control.createBuffer(state.length + 1);
@@ -504,72 +504,6 @@ namespace jacdac {
                 this.sensorState = SensorState.Stopping;
                 pauseUntil(() => this.sensorState == SensorState.Stopped);
             }
-        }
-    }
-
-    //% fixedInstances
-    export class SensorVirtualDriver extends Driver {
-        // virtual mode only
-        protected _localTime: number;
-        protected _lastState: Buffer;
-        private _stateChangedHandler: () => void;
-
-        constructor(name: string, deviceClass: number) {
-            super(name, DriverType.VirtualDriver, deviceClass);
-            this._lastState = control.createBuffer(0);
-            jacdac.addDriver(this);
-        }
-
-        public get state() {
-            return this._lastState;
-        }
-
-        /**
-         * Enables or disable streaming the sensor internal state
-         * @param on streaming enabled
-         */
-        //% blockid=jacdacsensorstreaming block="jacdac %sensor set streaming %on"
-        //% on.shadow=toggleOnOff weight=1
-        //% group="Input"
-        public setStreaming(on: boolean) {
-            const msg = control.createBuffer(1);
-            msg.setNumber(NumberFormat.UInt8LE, 0, on ? SensorCommand.StartStream : SensorCommand.StopStream);
-            this.sendPacket(msg);
-        }
-
-        public onStateChanged(handler: () => void) {
-            this._stateChangedHandler = handler;
-        }
-
-        public handlePacket(pkt: Buffer): boolean {
-            const packet = new JDPacket(pkt);
-            const command = packet.getNumber(NumberFormat.UInt8LE, 0);
-            this.log(`vpkt ${command}`)
-            switch (command) {
-                case SensorCommand.State:
-                    const state = packet.data.slice(1);
-                    const changed = !bufferEqual(this._lastState, state);
-                    const r = this.handleVirtualState(state);
-                    this._lastState = state;
-                    this._localTime = control.millis();
-                    if (changed && this._stateChangedHandler)
-                        this._stateChangedHandler();
-                    return r;
-                case SensorCommand.Event:
-                    const value = packet.data.getNumber(NumberFormat.UInt16LE, 1);
-                    control.raiseEvent(this.id, value);
-                    return true;
-                default:
-                    return this.handleCustomCommand(command, packet);
-            }
-        }
-
-        protected handleCustomCommand(command: number, pkt: JDPacket) {
-            return true;
-        }
-
-        protected handleVirtualState(state: Buffer) {
-            return true;
         }
     }
 }
